# swagger/swagger.yaml
...
info:
  contact: {}
paths:
...
  /update/{type}/{name}/{value}:
    post:
      description: Updates a single metric identified by type and name using a value
        from the URL
      parameters:
      - description: Metric type (counter or gauge)
        in: path
        name: type
        required: true
        type: string
      - description: Metric name
        in: path
        name: name
        required: true
        type: string
      - description: Metric value
        in: path
        name: value
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Metric updated
          schema:
            type: string
        "400":
          description: Invalid metric parameters or value
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update metric
      tags:
      - metrics
...
swagger: "2.0"